# Conflicts:
#	generator/src/main/java/fr/unice/polytech/creditrama/App.java

+++ clientService/src/main/java/fr/unice/polytech/si5/al/clientService/models/BankAccount.java
@@ -26,7 +26,7 @@ public class BankAccount {
     long clientId;
 
     @NotNull
-    int amount;
+    double amount;
 
     @Override
     public String toString() {
